It works!
It can now handle different concentrations/replicates for both compounds and controls. However, there is currently a bug with one of the assertions for early termination. I'm including pl-example23 to show that it is possible to handle combinations as compounds but that's not very efficient at the moment. Possibly some constraints are too weak or missing.