More tests! 🩺
More examples for testing. Some for testing the new model explicitly, such as diving an odd number of controls or plates with an odd number of rows (more fiddly than it sounds!). Also adapted some of the old examples (hence why the numbering is not sequential all of a sudden) to see how things are going.

This is still a baby model and it's not doing much.